+++ CHANGELOG.md
@@ -4,13 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [v1.3.0 (2022-01-12)](https://github.com/pestphp/pest/compare/v1.2.0...v1.3.0)
+## [v1.4.2 (2022-01-29)](https://github.com/nunomaduro/termwind/compare/v1.4.1...v1.4.2)
+### Fixed
+- `max-w` with `w-{fraction}` childs by @xiCO2k in https://github.com/nunomaduro/termwind/pull/118
+
+## [v1.4.1 (2022-01-29)](https://github.com/nunomaduro/termwind/compare/v1.4.0...v1.4.1)
+### Fixed
+- `<href>` with more width that the parent element by @xiCO2k in https://github.com/nunomaduro/termwind/pull/117
+
+## [v1.4.0 (2022-01-26)](https://github.com/nunomaduro/termwind/compare/v1.3.0...v1.4.0)
+### Added
+- Add support for `space-y` and `space-x`. by @xiCO2k in https://github.com/nunomaduro/termwind/pull/115
+
+## [v1.3.0 (2022-01-12)](https://github.com/nunomaduro/termwind/compare/v1.2.0...v1.3.0)
 ### Added
 - Add `ask()` method. by @xiCO2k in https://github.com/nunomaduro/termwind/pull/112
 - Add `max-w-{width}` class. by @xiCO2k in https://github.com/nunomaduro/termwind/pull/111
 - Add support for `py`, `pt` and `pb`. by @xiCO2k in https://github.com/nunomaduro/termwind/pull/113
 
-## [v1.2.0 (2021-12-16)](https://github.com/pestphp/pest/compare/v1.1.0...v1.2.0)
+## [v1.2.0 (2021-12-16)](https://github.com/nunomaduro/termwind/compare/v1.1.0...v1.2.0)
 ### Added
 - Add `terminal()->clear()` method. by @xiCO2k in https://github.com/nunomaduro/termwind/pull/107
 - Added support for HTML5 tags by @opuu in https://github.com/nunomaduro/termwind/pull/110
@@ -20,7 +32,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fix function_exists() checks by @mabar in https://github.com/nunomaduro/termwind/pull/108
 - Fix bug when using emojis with `w-full`. ([f2f426](https://github.com/nunomaduro/termwind/commit/f2f4261f9e2af1181a2816748fac7a6015316f8a))
 
-## [v1.1.0 (2021-12-06)](https://github.com/pestphp/pest/compare/v1.0.0...v1.1.0)
+## [v1.1.0 (2021-12-06)](https://github.com/nunomaduro/termwind/compare/v1.0.0...v1.1.0)
 ### Added
 - Support for Symfony 6 components ([67b9a9](https://github.com/nunomaduro/termwind/commit/67b9a9bab640dde70220d5b415378fe8be311a79))